requirements.txt
python-docx==1.2.0
PyPDF2==3.0.1
nltk==3.9.3
scikit-learn==1.8.0
fpdf2==2.8.7
jinja2==3.1.6
rake-nltk==1.0.6